# broker-service/docs/swagger.yaml
basePath: /
definitions:
  handlers.jsonResponse:
    properties:
      data: {}
      error:
        type: boolean
      message:
        type: string
    type: object
  handlers.requestPayload:
    properties:
      email:
        type: string
      password:
        type: string
      username:
        type: string
    type: object
host: localhost:8080
info:
  contact: {}
  title: kinosaya API
  version: "1.0"
paths:
  /:
    get:
      description: Returns a JSON response with success status.
      operationId: get-sample-response
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/handlers.jsonResponse'
      summary: hit the broker
  /auth/login:
    post:
      consumes:
      - application/json
      description: Logs in a user with the given data.
      parameters:
      - description: User data
        in: body
        name: requestPayload
        required: true
        schema:
          $ref: '#/definitions/handlers.requestPayload'
      produces:
      - application/json
      responses:
        "202":
          description: Successful registration
          schema:
            $ref: '#/definitions/handlers.jsonResponse'
        "401":
          description: Invalid credentials
          schema:
            $ref: '#/definitions/handlers.jsonResponse'
      summary: Logs a user
      tags:
      - Auth
  /auth/registration:
    post:
      consumes:
      - application/json
      description: Registers a new user with the specified data.
      parameters:
      - description: User data
        in: body
        name: requestPayload
        required: true
        schema:
          $ref: '#/definitions/handlers.requestPayload'
      produces:
      - application/json
      responses:
        "202":
          description: Successful registration
          schema:
            $ref: '#/definitions/handlers.jsonResponse'
        "401":
          description: Invalid credentials
          schema:
            $ref: '#/definitions/handlers.jsonResponse'
      summary: Register a new user
      tags:
      - Auth
swagger: "2.0"